resilent selected instrument retrieval
The code for retrieving the currently selected instrument was moved into the Hydrogen class as there were various copies of this code in the GUI. Some with proper locking of the audio engine, but most without.

Also, all instances the selected instrument is accessed are now checked for nullptr. There are various places selected instrument number is set and just one of these functions has to be off in order for Hydrogen to crash. (It did not happen at my end recently but it feels like a no brainer)