Instrument: store drumkit lookup
currently all instruments loaded in the current drumkit (of the song) are associated with one drumkit in the sound library. This is done by storing the drumkit name in the Instrument class and writing/reading them to/from a .h2song file. But just the name is not enough to uniquely determine the associated kit. From now on, also the drumkit lookup (where the drumkit is stored) is stored as well