SoundLibraryPanel: revisit highlighting
the drumkit last loaded into the song was highlighted in the SoundLibrary. However, there was not much information in this. For one the name is also shown as a label in the upper left corner of the PatternEditor. Also, each individual instrument in the instrument list of a song is associated with drumkit in the sound library. There is no "general" or "over all" drumkit. One could, for instance, load a drumkit a, drag-import instruments from drumkit b, and remove all original instruments from a. As a result a is still highlighted in the SoundLibraryPanel although it does not serve any purpose anymore.

With this patch the drumkit associated to the instrument currently selected (in the PatternEditor) is highlighted. Even when having multiple instruments of the same name, e.g. "Kick", this makes it still obvious where they came from and where to find there samples