Export song: license note
a mutable dialog window does inform the user about all licenses of samples loaded into the current song.

It does not check whether the instrument/sample is actually used. This would be both an overkill and lead to non-deterministic behavior.

The dialog also gives a warning in case a copyleft license or one requiring an attribution is present