JACK: relocate when transport is stopped
Due to limitations of a previous implementation the JACK driver did only process the incoming transport information in case transport was rolling within Hydrogen. The playhead would only be updated when starting transport and Hydrogen appeared to be out of sync with other JACK clients (see #1597).

Fortunately, with the new transport implementation of the audio engine I do not see any reason right now to still discard the transport information instead of performing a playhead update like all the other JACK clients do.