JACK: relocate in Pattern Editor
Whenever pattern mode was activated transport was reset to the beginning of the pattern in case a relocation was reported by the JACK server. However, this can very easily make Hydrogen and other clients go out of sync as well as forbids relocating the playhead using the ruler of the Pattern Editor.

Thats why transport position changes in pattern mode are now handled the same way as they are in song mode