Preferences: fix change propagation
The implementation of the Preferences::Changes propagation after closing the PreferencesDialog was pretty half-assed. It worked fine on fonts, colors, appearance tab, and some hand-picked other settings. But for most options in the General, MIDI, Audio, and OSC tab it was not implemented yet.

As of now the PreferencesDialog will check for changes in all options before attempting to write its values and broadcasts what has been changed by the user.

Also, there were some options that where not set in the Preferences::on_ok function but in individual callbacks, like per track audio output for JACK and the interpolation mode. Those were written even in casethe dialog was rejected.