PreferencesDialog: use current Preferences
Upon opening the PreferencesDialog did reload the local preferences of the user from file and used it to overwrite the current instance of the Preferences. This was done since the beginning of the versioning history.

I am not quite sure what is the purpose of this line but it overwrites all local changes to the Preferences that have not been saved yet (by regulary quitting Hydrogen). As a number of Buttons and options in the GUI alter settings in the Preferences this resets them to their previous state without updating them. The users changes are overwritten and the GUI and the core get slightly out of sync. I removed this line because it seem to have no purpose anymore.