AudioEngine: do not wrap after song size change
when creating a song with a couple of empty columns followed by one last pattern and toggling that last one while the playhead is in the empty ones, transport was wrapped regardless of whether loop mode was activated or not. Now, it only gets wrapped with loop mode activated. Else, the song is stopped and the playhead moves to the beginning of the song.