Buttons: click instead of press
As a legacy from the previous implementation on the `Button` widget many of them still used the `pressed()` signal for interaction instead of `clicked()` - which is the default one in Qt.

In this PR I ported all buttons to use `clicked()` instead. This means, play button etc. are not triggering their associated function upon pressing done the mouse button but, instead, on releasing it. This has the following benefits
- the UX of our custom buttons is consistent with the one of all other buttons used in dialogs etc.
- as we now base our `Button` widget on `QPushButton` it was possible to get the button and the associated core state out of sync. E.g. pressing the Sound Library button while the Instrument Editor is shown hides the latter and shows the former. But if the user - while still holding down the mouse button - moves the mouse pointer outside of the boundaries of the button and releases there, the QPushButton never gets clicked as this is Qts way of cancelling the clicking. Now, the InstrumentEditor button is still checked while the Sound Library is shown.