Remove LocalFileMng
the LocalFileMng class duplicated the code also found in src/core/Helpers/Xml but with a smaller set of write functions and different default argument and argument order for read functions. This made the code for reading/writing things from/to disk unnecessary complicated and harder to maintain.

Also, `SongWriter` and `SongReader` were removed and their code moved to `Song::writeTo` and `Song::loadFrom` in order to structure the Song class similar to other classes. Its loading is still a mess with a lot of duplications but all legacy parts of load drumkit pattern have been moved into the `Legacy` class. In addition, all code corresponding to the TinyXML support was moved to `Legacy` as well