fix license load and speed up instrument loading
Drumkits will now provide an overall license for each instrument instead loading a per-instrument one from disk

Each license associated with a drumkit path is now cached in a map located in the Hydrogen class. If there are multiple instruments from the same drumkit stored in a song, this speeds up song loading quite a bit.