Object: introduce CLOCK & CLOCKIN macros
The CLOCK macro introduced for more easy profiling of member functions. It can be invoked without or with a string argument in a member function of a class inheriting from `Object`.

The first time CLOCK() is called it registers the current time. Each further call will calculate the time difference since the previous call of CLOCK() and prints it in the DEBUGLOG. The previous time used to calculate the time difference is a static variable of a particular class and allows for clocking to be used across different member functions.

In case just the time difference of a single function is desired CLOCKIN() can be used at its beginning to reset the internal clocking. (This assumes that there are no parallel threads accessing the same singleton which is usually fullfilled in Hydrogen).