reintroduce asserts
while reworking load_from/save_to functions I replaced the `assert`s in the code by if + log + return statement in order to make the execution of the function save to use. But I agree asserts are a nice thing as an early warning - especially in case they are triggered during the unit tests. So, I introduced them again next to the if statements