Fix post-deallocation access of HydrogenApp
In deallocation, the lifetime of the HydrogenApp object (which is
allocated and deallocated in MainForm, which feels a bit upside-down).
EventListener objects may outlive the end of the lifetime of the
HydrogenApp object and will be deallocated as MainForm and its
children are destroyed by the framework.

This change makes the deallocation of some EventListeners (those
which explicitly remove themselves from HydrogenApp's client list)
robust against potential non-existence of their event source.