reference drumkit in song using name and path
`Song::m_sCurrentDrumkitName` was replaced by `Song::m_sLastLoadedDrumkit` to emphasis the intended meaning of the variable and `Song::m_currentDrumkitLookup` was dropped in favour for `Song::m_sLastLoadedDrumkitPath`. Both new variables are now written and read from disk.

The GUI was mostly freed of remaining `Filesystem::Lookup` but not adopted to use path instead. So, temporarily drumkit loading will be based just on the drumkit name again and the highlighting of the current drumkit/instrument in the SoundLibrary is disabled.