Instrument: write drumkit name
In addition to "drumkitPath" the previously used "drumkit" (name of the corresponding) drumkit is written to disk again. This ensures both backward compatibility of the created song was well as portability. While absolute paths are nice identifiers of a drumkit locally, they are not helpful when transferring the song to a different device. In such a case, the name of the drumkit will be used for a classic drumkit serach.