move SoundLibraryDatabase into core
Caching the drumkit licenses proofed quite efficient during startup. As there are probably a number of different use cases for caching, the SoundLibrary requires some work to be fixed anyway, and holding a list of all available songs, patterns, and drumkits in memory is more of a core than a gui thing to do, `gui/src/SoundLibrary/SoundLibraryDataStructures` has been moved to `core/SoundLibrary/SoundLibraryDatabase` and `core/SoundLibrary/SoundLibraryInfo`.

Note that the current version of the database does only cover patterns and neither drumkits nor songs. Seems as this is a stub which had not been finished.