Drumkit: streamline drumkit load
The possible functions to load a Drumkit from disk has been reduced to Drumkit::load() which is supposed to be not used directly anymore but, instead, is called from SoundLibraryDatabase.

For retrieving a pointer to a specific drumkit one should use SoundLibraryDatabase::getDrumkit() which is able to work with both drumkit names as well as relative and absolute paths. If the drumkit corresponding to the supplied name/path was not loaded yet, it is done so internally and added to the cache.

For applying a drumkit to a Song one should use CoreActionController::setDrumkit which comes is two versions. One for drumkit names/paths wrapping another one for shared pointer to a kit. Song::loadDrumkit has been renamed to Song::setDrumkit and Hydrogen::loadDrumkit had been dropped in favor for the version found in CoreActionController to make more obvious which is the one to pick.

SoundLibraryDatabase is now able to handle drumkits which are neither present in the user or system drumkit folder and are handed to Hydrogen via CLI argument or OSC action.