drop SoundLibrarySaveDialog
it has been an almost exact duplicate of SoundLibraryPropertiesDialog that was slowly getting out of sync. Now, only the latter class will be used.

If opened to alter properties, the name of the drumkit can not be altered anymore. In order to create a new drumkit from an existing one by altering its name, one has to use MainForm > Drumkits > Save As instead.