SoundLibraryExportDialog: rework
The combo box for choosing a drumkit was removed and the dialog only acts on the one it was called on. It exports either a stack drumkit in case it was invoked via right click memu in the SoundLibrary or the last loaded drumkit with the instrument and component list of the current song if invoked from the main menu.