Rework status message handling
All code covering the handling of status messages in the `PlayerControl` has been moved to the dedicated class `StatusMessageDisplay`. There is now one timeout for all messages and a single public function to show a message. The class checks the length of the message and decides internally whether to scroll the message or display it whole.

Clicking the status message widget opens a dropdown (QMenu) displaying the last 100 status messages.

`StatusMessageDisplay::showMessage` features a second argument using which the calling context can be identified. Calling it several times in a row using the same argument results in only the last message to be added to the history. Else, status messages for e.g. setting the strip volume would flood the history.

A couple of status messages have been reworked in order make them more verbose and unique (e.g. by adding the name of the instrument line)