Instrument: fix drumkit load using lookup
There was a corner case I did not had in mind when rewriting the code used for drumkit loading (but luckily encountered when trying to update the images of the manual).

It is possible the song was composed with a custom version of a
system-level drumkit stored in user space. When loaded again in a
fresh installed Hydrogen the custom user-level one will not be
present anymore but it is plausible to fall back to the
system-level one. (The other way around is also possible but much
more unlikely. Nevertheless we will use the stacked search in one
final effort)