Shotlist: trigger shots using core event
A number of GUI slots have been changed in v1.2 to just be wrappers around public methods of `CoreActionController`. While this reduces redundancy as well as complexity of the GUI by a margin it also means that triggering the slot for activating e.g. song mode does not put the GUI in song mode right away. The core will be set to song mode but GUI will be informed and subsequently updated using an event in `EventQueue`.

This is a problem when using the Shotlist as it just processes one line after another and expects and does not wait for the GUI to be synced with the core again.

To fix this problem, the next shot is now triggered using an event of the `EventQueue` too. Since queued events are processed sequentially this slows down the pace of the shotlist processing to match the update of the GUI.

Processing a `Shotlist` becomes a lot slower using this patch but I think this is not relevant since it is just invoked by a developer like once a year.