Minimize updated areas of editors for playhead movement
While playing, the song editor and pattern editors will update
periodically to show the progress of the playhead across the song
or pattern. The backgrounds are cached as pixmaps so this is a lot
faster than redrawing the whole song or pattern, but blitting the pixmap can still
take significant amounts of time when large areas of screen space
are taken up by these widgets (particularly on, say, a Raspberry
Pi attached to a 2560x1400 display). It's unclear if Qt is using
any hardware acceleration for this, but I've seen large amounts of
CPU time used in Linux and macOS profiling.

This change modifies the timer-driven updates to call update() on
only the minimal area necessary to correctly keep the widget fresh,
and erase the old playhead and render the new one.