NSM: link drumkit on song save
Previously, the drumkit was linked into the session folder the moment a new song was loaded into the session or a drumkit was loaded. This, however, did allow for the song file and the linked drumkit to get out of sync when Hydrogen was closed without saving the song. To prevent this, drumkit linking is now delayed till song save.