Reworking NSM state management
The session (NSM) support of Hydrogen is designed to allow the application to
retrieve all files required to make a song reproducible from a session folder.
This includes either linking a drumkit from the data folder or supporting a
local drumkit directory (in order to make sessions portable).

The above mentioned feature was updated to work with the newly introduced
`SoundLibraryDatabase`. In addition, the latter got a new function called
`updateDrumkit` for a more efficient update of just a single kit corresponding
to a provided path. It can also be used to load a new kit into the database.

The previous implementation also showed some inconsistencies that were fixed too.
Whenever a drumkit or a new song with a different kit was loaded, the drumkit was
immediately linked into the session folder. If, however, the user decided to discard
those changes by closing the session without changing or an e.g. power outage occured,
the song in the session got into a bricked state with new kit already linked but
the .h2song file still containing references to the old one. This was solved by
introducing a transient state into the session support. Whenever a different song or
drumkit was loaded `Song::m_lastLoadedDrumkitPath` as well as the kit paths in the
instruments were set to the ones corresponding to the drumkits in the data folder.
This way the new kit could be used with the old one still linked into the session
folder. In addition, a flag will be set that triggers relinking and the usage of
session paths for the drumkit as soon as the session is saved. This way song and drumkit
are always in sync and changes like song or drumkit importing can be reverted by
aborting the session.

New session are adapted to the new concept of transient empty songs. Both song and
drumkit will be stored/linked into the session folder at the first explicit save action
of the user. To indicate that this is required, new songs under session management
are always flagged modified.

(The autosave function does not trigger drumkit relinking)