support relative drumkit/sample paths in NSM session
Since the introduction of the `SoundLibraryDatabase` paths to drumkit folders are used as unique identifiers for drumkits. This, however, causes a number of problems in the NSM implementation. Absolute paths not only prevent the session from being portable, they also brick the .h2song in case the session gets renamed or duplicated. To avoid this, we now support drumkit and sample paths relative to the session folder.