TransportTests: fix note queue check for extreme parameters
if the setting for buffer size and sample rate differ significantly from the ones usually used in the unit tests, e.g. buffer size 5000 and sample rate 1024 - an artificial example, I know, but it helps to spot bugs - set in TestHelpers::varyAudioDriverConfig(9), the song note queue will be flooded with notes.

The humanization was ramped up and the velocities of the individual kick notes in the double bass part were diversified because AudioEngine::testCheckAudioConsistency() was not able to keep the enqueued notes appart.

In addition, `AudioEngine::processAudio()` was called within `AudioEngine::testToggleAndCheckConsistency()` after copying the notes from the note queue of the audio engine. This, however, is not right since the former can/will pop notes from the queue.