AudioEngine: test and fix column change
Apart from checking the integrity of transport position and computed tick interval in `AudioEngine::testSongSizeChange()` the consistency of the column transport resides in is checked as well. After all, glitches in this variable will result in jumps of the playhead and are annoying for the user.

This uncovered an edge case not properly handled in `AudioEngine::updateSongSize()` yet. Whenever the current column before toggling is beyond the end of the song after toggling transport is loop to a particular position within the song. E.g. there are several empty patterns in front of a final one, playhead resides in one of the empty ones, and the users deactivates the final pattern. From the user perspective, however, it is not obvious why transport is not just wrapped to the beginning of the song but to somewhere within it. This was now fixed.

In addition, a matching call to `AudioEngine::incrementTransportPosition` was missing in `AudioEngine::testSongSizeChange` was missing causing pattern and column related variables to go out of sync (did not affect the test results previously).