AudioEngine: fix tick mismatch near tempo markers
whenever transport was located to a tick very, very close _before_ a tempo marker `AudioEngine::m_fTickMismatch` could be calculated wrong. It happened when the final frame in double calculated for the provided tick in `AudioEngine::computeFrameFromTick` was smaller than the position of the next tempo marker in frames but its rounded equivalent was located after the tempo marker. In this case the calculation of the mismatch requires to take both the tick size before and the one after the marker into account, which was not done before.

This is an edge case that could only occur when using the JACK server, would be extremely rare, would only affect transport if the user is changing the size of the song in that exact moment, and would then only result in an unnoticable small glitch in the transport position. But it got caught in the tests and deserved fixing :)