Introduce TransportPosition
first draft of extending the former TransportInfo into class encapsulating a transport position. Both the actual position of the transport and the one of the playhead (ahead of the former by a lookahead) are now kept alongside of each other and all parts of Hydrogen accessing a position have to explicitly state which one to interact with. This makes the code (especially the one of the AudioEngine) much more easy to grasp.