rename g/setFrames() -> g/setFrame()
Formerly the current transport position in ticks was called m_tick and the corresponding one in frames m_frames. As a result most functions handling the frame position carried the plural till today. This was a little nuissence and resulted in quite a number of typos over the years and since I already touch transport position definitions, I will straighten this one up and only use singular in all places.