AudioEngine: consistent playhead update on tempo change
The previous design of `AudioEngine::computeTickInterval` - use to determine the next range of ticks covered in `AudioEngine::updateNoteQueue` - had an conceptional flaw. Whenever there is a tempo change the current tick position is conserved and the current frame position has to be updated. Since transport processing in the audio engine - and in updateNoteQueue - relies on frames, the tick interval covered will not align with the previous one. In order to prevent holes, the previous implementation remembered the last tick interval end and either consumed all the space between the previous (smaller) interval end and the start of the current one by making a rather huge jump/interval or waited till transport reached the end of the previous (larger) interval. This ensured consistency in the covered ticks, prevented notes from being missed, and kept the transport position consistent. However, now that we have a separate `AudioEngine::m_pPlayheadPosition`, which we do also cover in the unit tests, it quickly became apparent that the design above results in jumps or short stops of the playhead in the GUI. This is no good at all.

The new take of the implementation also remembers the last tick interval end but calculates a tick offset between it and the new one on each tempo change. By applying the offset (in ticks) on each subsequent calculation in `AudioEngine::computeTickInterval` we can ensure the playhead smoothly moves on with a different speed when encountering a tempo change.

When fixing the issue above another problem unearthed in the calculation of the tick interval. The lookahead is not consistent for constant tempo. This is due to rounding errors when converting its tick component into frame even though the latter is perfectly fine. This lookahead fluctuation by +/-1 frame is catched by remembering the lookahead, applying the stored one in case of such an legit mismatch, and resetting it in case of a tempo change or relocation. This is also covered within an unit test as the consistency of the lookahead is essential for a consistency of the calculated tick intervals