Lock AudioEngine while calling setNextBpm
When changing tempo while transport was rolling with the Timeline deactivated there were occassional glitches of the playhead position (and note rendering). This was due to calls to `AudioEngine::setNextBpm` without the audio engine being locked. Sometimes the update happened during audioEngine_process resulting in an inconsistent tempo throughout the run. This can lead to a transport position begin already updated to the new speed while playhead remains the same.

This was fixed by placing locks around each call of `AudioEngine::setNextBpm`.