SongEditor: coarse-grain playback track volume
till now the volume of the playback track adjusted via the corresponding fader was set in float precision (while only shown up to the second digit after point as a tooltip). This caused me some headache while creating an unit test. I wanted to set the volume to 1 and expected the playback track to pass through as is. But, instead, the volume was set to 0.985674... which does not seem right.

With this patch we can only set the playback volume with a precision up to the second digit after point and the value displayed in the tooltip will also be written to the song file. If anyone strives for a more precise volume, she can still edit the song file manually.

In addition, there is now a status bar message indicating the new playback track volume value.