test: check consistency of playback track
a test was added which plays back just the playback track and checks whether the result matches the playback track used as input.

as a prerequisite to allow such a test the playback track can now also be loaded relative to the song file. Its absolute path written by Hydrogen must, however, be manually edited. This is a precaution to prevent bricked songs on user level while at the same till allow for the playback track to be tested.