tests: audio file assertion with tailing silence
As it is a bit of a nuissance to adjust incoming samples/playback tracks in the unit tests in such a way they are identical to the exported audio file, I did introduce another assertion for audio file data that eases things. With `H2TEST_ASSERT_AUDIO_FILES_DATA_EQUAL(expected, actual)` the `actual` - the exported audio file - must identical to expected but is allowed to be longer if all additional values are 0. This way the unit test do not fail on tailing silence anymore