tests: exception handling + noteEnqueuingTimeline stub
The signal path of the AudioEngine tests was flawed. As output intended for the console without any verbosity level activated was issued directly using `qDebug()` and the log queue was only flushed after the test on `TransportTest::tearDown` the error message gets lost in a ton of log message quite frequently.

Instead, the `AudioEngineTests` do now propagate failures using exceptions, which are caught in `TransportTest` and rethrown as CppUnits default exceptions. This way athe error message comes last and is formatted the same way as in all the other unit tests.

Also, a stub for a new test concerning the consistency of note placement/enqueuing while using tempo markers is introduced