AudioEngine: use PlayheadPosition only internally
some parts of the GUI and core did access and use the playhead position (about to be renamed) instead of the transport position. Upon revisiting, this is actually wrong. The playhead position only used to add notes to the song queue and to update the playing patterns in order to select new notes. All other transport related things are handled by the transport position.

There is still a conceptual flaw as the playing patterns are updated by the playhead position (transport + lookahead) and thus the playing patterns indicators in the GUI will change prior to transport reaching the corresponding column