AudioEngine: fix bug in setSong
in `AudioEngine::setSong` just the first pattern in the first column added to `AudioEngine::m_pPlayingPatterns` using the assumption that a newly loaded song is in selected pattern mode. But this is not true. If the song was saved in song mode, it will be set to song mode immediately after loading. If multiple patterns are present in the first column, Hydrogen will only play the first one instead of all of them.