AudioEngine: fix lookahead treatment in computeTickInterval
when passing a tempo marker the lookahead is constantly changing and needs to be fixed to the previous value. Else, we would risk to loose notes due to holes in the covered ticks in updateNoteQueue

`AudioEngine::m_nLastLeadLagFactor` was moved into the `TransportPosition` class. Setting it when calling `updateBpmAndTickSize` for both transport and queuing position caused bugs and strange behavior.