PlayerControl: more fluent wheel event on BPM spin box
`PlayerControl::updatePlayerControl` does not update the BPM widget anymore in case the user hovers it. This prevents it from picking up old values while using e.g. the mouse wheel to increase the value.

It is save since tempo changes happening in the core are triggering EVENT_TEMPO_CHANGE and PlayerControl response to it. It does not rely on the timer based update.

Setting the tempo using wheel event for the BPM widget is still a bit rough. This is because changes are not picked up instantly by the core but in the next processing cycle and thus a tempo change event is triggered with a little delay and causes the widget to be set to another value. Unfortunately, QDoubleSpin box has not different signal to distinguish between user and scripting value changes. But I think we can leave it like this for now.