AudioEngine: fix song end detection on relocation
There were some false positive detections of song end in `AudioEngine::updateNoteQueue` due to an unsuited application of offsets and rouding errors. This is now also covered in an unit test.

`AudioEngine::calculateTransportOffsetOnTempoChange` does not calculate offsets anymore in case transport did not roll yet after a relocation or stop.