AudioEngine: fix updateSongSize handling
Previously, whenever the current column exceeds the number of columns after the song size change transport was relocated to the beginning. This, however, prooved to be cumbersome when working with JACK and additional applications. Toggling grid cells in the song editor would alter transport in all other application which is no good. Also, transport is now consistent when changing song size and restoring the old size.