Sampler: fix note end calculation
The calculation of nNoteEnd in the sample rendering (used to determine the release frame in applyADSR) is off. When performing resampling it does not take the initial silence at the beginning of the buffer into account. E.g. I have a buffer of 4096 frames and a metronome sample of 2700 frames placed by the audio engine in such a way it starts at frame 3520 within the buffer (second click in the audio sample attached above). Since this initial silence was missed, nNoteEnd is set to 2071 ( < 3520 ) and applyADSR goes right into the release state and the faint metronome click we hear is actually the default 1000 release frames of the original sample multiplied with the release envelope.

This problem was not caught by an unit test since it uses a buffer size of 1024 and there seem to be no samples involved which are smaller than that.