AudioEngine: fix velocity humanization
Whenever velocity humanization was used, a constant offset was introduced. As a result adding a tiny amount of velocity humanization caused the velocity of all notes to increase in addition to introducing some small jitter of their values.

This can not be right. Especially since there is no offset in any of the other humanization methods. Consulting git logs did not yield a clue either as this part of the code predates the migration to git.

I removed the offset and left clamping the new velocity value to be done inside `Node::set_velocity`