Note: do not incorporate instrument pitch offset
the pitch of a note consists of three parts: a random component, `Note::__pitch` itself (barely used. E.g. within the metronome), and the pitch offset of the associated instrument. Previously, the last component was manually inserted in most places right before handing the note to the `Sampler. Most places, because it was missed in instrument preview, sample playback in the SampleEditor, and when previewing a certain velocity in the InstrumentEditor > Layer view.

As all occurrences exclusively provided the pitch offset of the instrument associated to a note, the whole thing was into the Sampler were the calculation of the total pitch does now include the pitch offset of the instrument via a tweak of `Note::get_total_pitch`. This assures the instrument pitch offset is always respected.