AudioEngine: update metronome volume on demand
Previously, the value of the metronome volume set within the Preferences was used to update the volume of the metronome instrument in the audio engine on each tick passed.

Now, it is assigned on startup, if another preference is loaded, or there were changes applied in the Audio tab of the PreferencesDialog by the user.