tests: CLI option to write log to file + Logger fix
the `tests` binary has now an additional command line option `-o` or `--output-file`. If set, the log level will be Info per default, all output will be directed to the provided file, and _no_ output will be handed over to stdout.

The provided file can be either a filename or a relative as well as absolute path.

In addition, the `Logger` class is now properly destructed in `tests` in order to assure the log files content is properly flushed and the file descriptor is closed.