AppVeyor: upload output of failing tests
AppVeyor is a great help to check whether build and tests do pass on all supported platforms. For debugging failing test, however, it is not very helpful as just the plain error message causing the failure is printed in the console.

With this patch all logs (verbosity level INFO) of the tests are written to a file, which is only uploaded in case the tests failed. Console output is not affected.

(It is more of a hack as it was intended to be but I could not get the Windows script to proceed after the tests binary returned a non null exit code. So, Linux and macOS scripts upload the test output on failure while Windows scripts delete them on success and a later stage uploades them in case they are still around)