DiskWriterDriver: fix progress calculation
In Windows 32bit builds calculation for the `EVENT_PROGRESS` `(int) ((float)X / (float)X * 100.0)` equates to 99 when exporting audio in some unit tests. This causes the tests to wait forever and the Windows 32bit pipeline to fail.

Now, a ceiling is used in the combination with the integer conversion which fixes the error.