AudioEngine: Windows 32bit lax TS comparison
For some reason two identical numbers (according to their values
when printing them) are not equal to each other in 32bit
Windows. This results in a thorough update of the AudioEngine state in
`updateBpmAndTickSize` even when there are no tempo changes
encountered and thus the lookahead will be recalculated. These
recalculations can result in a small mismatch (+/-1) and tiny
inconsistencies. And it makes a totally fine unit test fail.

Course graining the tick change in here will do no harm
except of for preventing tiny tempo changes. Integer value
changes should not be affected.

I do not think this should be the end of the line for this problem but
it is probably a good idea to postpone it after 1.2-beta.