update License notice
in some files the license note has been normalized and missing ones were added.

In a couple of instances years covered were changed from the actual year of creation, e.g. 2021, to the range of years 2008-2022. This is less clean up helps to do a search and replace for all the license notices next year