AppVeyor: install JACK via MSYS2
and obtain `libssl` and `libcrypto` from the preinstalled MSYS2 `openssl` package. (Even in case it would not be present in future versions of AppVeyor it is a depedency of `qt5` and would be installed anyway).

This allows us to get rid of all the Chocolatey installs which prooved to be less reliable and routinely breaking the Windows build pipeline. (JACK was only added during the recent years to MSYS2 pacman repos and was not present when our AppVeyor script was written).

Fixes #1704